requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
flask-cors==4.0.0
gunicorn==21.2.0
neo4j==5.15.0
